+++ R/Estimator.R
@@ -27,8 +27,7 @@
 #' @param batchSize batchSize to use
 #' @param epochs  how many epochs to train for
 #' @param device  what device to train on, can be a string or a function to
-#' that evaluates
-#' to the device during runtime
+#' that evaluates to the device during runtime
 #' @param optimizer which optimizer to use
 #' @param scheduler which learning rate scheduler to use
 #' @param criterion loss function to use
@@ -40,26 +39,32 @@
 #' `name`,
 #' `fun` needs to be a function that takes in prediction and labels and
 #' outputs a score.
+#' @param accumulationSteps how many steps to accumulate gradients before
+#' updating weights, can also be a function that is evaluated during runtime
 #' @param seed seed to initialize weights of model with
 #' @export
-setEstimator <- function(learningRate = "auto",
-  weightDecay = 0.0,
-  batchSize = 512,
-  epochs = 30,
-  device = "cpu",
-  optimizer = torch$optim$AdamW,
-  scheduler = list(fun = torch$optim$lr_scheduler$ReduceLROnPlateau,
-                   params = list(patience = 1)),
-  criterion = torch$nn$BCEWithLogitsLoss,
-  earlyStopping = list(useEarlyStopping = TRUE,
-                       params = list(patience = 4)),
-  metric = "auc",
-  seed = NULL
-) {
-
+setEstimator <- function(
+    learningRate = "auto",
+    weightDecay = 0.0,
+    batchSize = 512,
+    epochs = 30,
+    device = "cpu",
+    optimizer = torch$optim$AdamW,
+    scheduler = list(
+      fun = torch$optim$lr_scheduler$ReduceLROnPlateau,
+      params = list(patience = 1)
+    ),
+    criterion = torch$nn$BCEWithLogitsLoss,
+    earlyStopping = list(
+      useEarlyStopping = TRUE,
+      params = list(patience = 4)
+    ),
+    metric = "auc",
+    accumulationSteps = NULL,
+    seed = NULL) {
   checkIsClass(learningRate, c("numeric", "character"))
   if (inherits(learningRate, "character") && learningRate != "auto") {
-    stop(paste0('Learning rate should be either a numeric or "auto", 
+    stop(paste0('Learning rate should be either a numeric or "auto",
                 you provided: ', learningRate))
   }
   checkIsClass(weightDecay, "numeric")
@@ -71,7 +76,14 @@ setEstimator <- function(learningRate = "auto",
   checkIsClass(earlyStopping, c("list", "NULL"))
   checkIsClass(metric, c("character", "list"))
   checkIsClass(seed, c("numeric", "integer", "NULL"))
-
+  
+  if (!is.null(accumulationSteps) && !is.function(accumulationSteps)) {
+    checkHigher(accumulationSteps, 0)
+    checkIsClass(accumulationSteps, c("numeric", "integer"))
+    if (batchSize %% accumulationSteps != 0) {
+      stop("Batch size should be divisible by accumulation steps")
+    }
+  }
 
   if (length(learningRate) == 1 && learningRate == "auto") {
     findLR <- TRUE
@@ -81,34 +93,52 @@ setEstimator <- function(learningRate = "auto",
   if (is.null(seed)) {
     seed <- as.integer(sample(1e5, 1))
   }
-  estimatorSettings <- list(learningRate = learningRate,
-                            weightDecay = weightDecay,
-                            batchSize = batchSize,
-                            epochs = epochs,
-                            device = device,
-                            earlyStopping = earlyStopping,
-                            findLR = findLR,
-                            metric = metric,
-                            seed = seed[1])
+  estimatorSettings <- list(
+    learningRate = learningRate,
+    weightDecay = weightDecay,
+    batchSize = batchSize,
+    epochs = epochs,
+    device = device,
+    earlyStopping = earlyStopping,
+    findLR = findLR,
+    metric = metric,
+    accumulationSteps = accumulationSteps,
+    seed = seed[1]
+  )
 
   optimizer <- rlang::enquo(optimizer)
   estimatorSettings$optimizer <- function() rlang::eval_tidy(optimizer)
-  class(estimatorSettings$optimizer) <- c("delayed",
-                                          class(estimatorSettings$optimizer))
+  class(estimatorSettings$optimizer) <- c(
+    "delayed",
+    class(estimatorSettings$optimizer)
+  )
 
   criterion <- rlang::enquo(criterion)
   estimatorSettings$criterion <- function() rlang::eval_tidy(criterion)
-  class(estimatorSettings$criterion) <- c("delayed",
-                                          class(estimatorSettings$criterion))
+  class(estimatorSettings$criterion) <- c(
+    "delayed",
+    class(estimatorSettings$criterion)
+  )
 
   scheduler <- rlang::enquo(scheduler)
   estimatorSettings$scheduler <- function() rlang::eval_tidy(scheduler)
-  class(estimatorSettings$scheduler) <- c("delayed",
-                                          class(estimatorSettings$scheduler))
+  class(estimatorSettings$scheduler) <- c(
+    "delayed",
+    class(estimatorSettings$scheduler)
+  )
 
   if (is.function(device)) {
-    class(estimatorSettings$device) <- c("delayed",
-                                         class(estimatorSettings$device))
+    class(estimatorSettings$device) <- c(
+      "delayed",
+      class(estimatorSettings$device)
+    )
+  }
+  
+  if (is.function(accumulationSteps)) {
+    class(estimatorSettings$accumulationSteps) <- c(
+      "delayed",
+      class(estimatorSettings$accumulationSteps)
+    )
   }
 
   estimatorSettings$paramsToTune <- extractParamsToTune(estimatorSettings)
@@ -133,19 +163,29 @@ fitEstimator <- function(trainData,
                          analysisPath,
                          ...) {
   start <- Sys.time()
-
-  # check covariate data
-  if (!FeatureExtraction::isCovariateData(trainData$covariateData)) {
-    stop("Needs correct covariateData")
-  }
-
   if (!is.null(trainData$folds)) {
     trainData$labels <- merge(trainData$labels, trainData$fold, by = "rowId")
   }
-  mappedCovariateData <- PatientLevelPrediction::MapIds(
-    covariateData = trainData$covariateData,
-    cohort = trainData$labels
-  )
+
+  if (modelSettings$modelType == "Finetuner") {
+    # make sure to use same mapping from covariateIds to columns if finetuning
+    path <- modelSettings$param[[1]]$modelPath
+    oldCovImportance <- utils::read.csv(file.path(path,
+                                                  "covariateImportance.csv"))
+    mapping <- oldCovImportance %>% dplyr::select("columnId", "covariateId")
+    numericalIndex <- which(oldCovImportance %>% dplyr::pull("isNumeric"))
+    mappedCovariateData <- PatientLevelPrediction::MapIds(
+      covariateData = trainData$covariateData,
+      cohort = trainData$labels,
+      mapping = mapping
+    )
+    mappedCovariateData$numericalIndex <- as.data.frame(numericalIndex)
+  } else {
+    mappedCovariateData <- PatientLevelPrediction::MapIds(
+      covariateData = trainData$covariateData,
+      cohort = trainData$labels
+    )
+  }
 
   covariateRef <- mappedCovariateData$covariateRef
 
@@ -161,13 +201,16 @@ fitEstimator <- function(trainData,
     )
   )
 
-  hyperSummary <- do.call(rbind, lapply(cvResult$paramGridSearch,
-                                        function(x) x$hyperSummary))
+  hyperSummary <- do.call(rbind, lapply(
+    cvResult$paramGridSearch,
+    function(x) x$hyperSummary
+  ))
   prediction <- cvResult$prediction
   incs <- rep(1, covariateRef %>%
-                dplyr::tally() %>%
-                dplyr::collect() %>%
-                as.integer())
+      dplyr::tally() %>%
+      dplyr::collect() %>%
+      as.integer()
+  )
   covariateRef <- covariateRef %>%
     dplyr::arrange("columnId") %>%
     dplyr::collect() %>%
@@ -180,26 +223,35 @@ fitEstimator <- function(trainData,
   comp <- start - Sys.time()
   result <- list(
     model = cvResult$estimator,
-
     preprocessing = list(
-      featureEngineering = attr(trainData$covariateData,
-                                "metaData")$featureEngineering,
-      tidyCovariates = attr(trainData$covariateData,
-                            "metaData")$tidyCovariateDataSettings,
+      featureEngineering = attr(
+        trainData$covariateData,
+        "metaData"
+      )$featureEngineering,
+      tidyCovariates = attr(
+        trainData$covariateData,
+        "metaData"
+      )$tidyCovariateDataSettings,
       requireDenseMatrix = FALSE
     ),
     prediction = prediction,
     modelDesign = PatientLevelPrediction::createModelDesign(
       targetId = attr(trainData, "metaData")$targetId,
       outcomeId = attr(trainData, "metaData")$outcomeId,
-      restrictPlpDataSettings = attr(trainData,
-                                     "metaData")$restrictPlpDataSettings,
+      restrictPlpDataSettings = attr(
+        trainData,
+        "metaData"
+      )$restrictPlpDataSettings,
       covariateSettings = attr(trainData, "metaData")$covariateSettings,
       populationSettings = attr(trainData, "metaData")$populationSettings,
-      featureEngineeringSettings = attr(trainData$covariateData,
-                                        "metaData")$featureEngineeringSettings,
-      preprocessSettings = attr(trainData$covariateData,
-                                "metaData")$preprocessSettings,
+      featureEngineeringSettings = attr(
+        trainData$covariateData,
+        "metaData"
+      )$featureEngineeringSettings,
+      preprocessSettings = attr(
+        trainData$covariateData,
+        "metaData"
+      )$preprocessSettings,
       modelSettings = modelSettings,
       splitSettings = attr(trainData, "metaData")$splitSettings,
       sampleSettings = attr(trainData, "metaData")$sampleSettings
@@ -219,7 +271,8 @@ fitEstimator <- function(trainData,
   )
 
   class(result) <- "plpModel"
-  attr(result, "predictionFunction") <- "predictDeepEstimator"
+  attr(result, "predictionFunction") <-
+    "DeepPatientLevelPrediction::predictDeepEstimator"
   attr(result, "modelType") <- "binary"
   attr(result, "saveType") <- modelSettings$saveType
 
@@ -255,13 +308,20 @@ predictDeepEstimator <- function(plpModel,
   # get predictions
   prediction <- cohort
   if (is.character(plpModel$model)) {
-    modelSettings <- plpModel$modelDesign$modelSettings
     model <- torch$load(file.path(plpModel$model,
-                                  "DeepEstimatorModel.pt"),
+                                  "DeepEstimatorModel.pt"),                        
                         map_location = "cpu")
-    estimator <- createEstimator(modelType = modelSettings$modelType,
-                                 modelParameters = model$model_parameters,
-                                 estimatorSettings = model$estimator_settings)
+    if (is.null(model$model_parameters$model_type)) {
+      # for backwards compatibility
+      model$model_parameters$model_type <- plpModel$modelDesign$modelSettings$modelType
+    }
+    model$estimator_settings$device <-
+      plpModel$modelDesign$modelSettings$estimatorSettings$device
+    estimator <-
+      createEstimator(modelParameters =
+                      snakeCaseToCamelCaseNames(model$model_parameters),
+                      estimatorSettings =
+                      snakeCaseToCamelCaseNames(model$estimator_settings))
     estimator$model$load_state_dict(model$model_state_dict)
     prediction$value <- estimator$predict_proba(data)
   } else {
@@ -292,152 +352,81 @@ gridCvDeep <- function(mappedData,
                        modelLocation,
                        analysisPath) {
   ParallelLogger::logInfo(paste0("Running hyperparameter search for ",
-                                 modelSettings$modelType, " model"))
-
+                                 modelSettings$modelType,
+                                 " model"))
   ###########################################################################
 
   paramSearch <- modelSettings$param
 
-  # TODO below chunk should be in a setupCache function
-  trainCache <- trainingCache$new(analysisPath)
-  if (trainCache$isParamGridIdentical(paramSearch)) {
-    gridSearchPredictons <- trainCache$getGridSearchPredictions()
-  } else {
-    gridSearchPredictons <- list()
-    length(gridSearchPredictons) <- length(paramSearch)
-    trainCache$saveGridSearchPredictions(gridSearchPredictons)
-    trainCache$saveModelParams(paramSearch)
-  }
+  # setup cache for hyperparameterResults
+  trainCache <- setupCache(analysisPath, paramSearch)
+  hyperparameterResults <- trainCache$getGridSearchPredictions()
 
   dataset <- createDataset(data = mappedData, labels = labels)
 
-  fitParams <- names(paramSearch[[1]])[grepl("^estimator",
-                                             names(paramSearch[[1]]))]
-  findLR <- modelSettings$estimatorSettings$findLR
   if (!trainCache$isFull()) {
     for (gridId in trainCache$getLastGridSearchIndex():length(paramSearch)) {
-      ParallelLogger::logInfo(paste0("Running hyperparameter combination no ",
-                                     gridId))
+      ParallelLogger::logInfo(paste0(
+        "Running hyperparameter combination no ",
+        gridId
+      ))
       ParallelLogger::logInfo(paste0("HyperParameters: "))
       ParallelLogger::logInfo(paste(names(paramSearch[[gridId]]),
-                                    paramSearch[[gridId]], collapse = " | "))
-      currentModelParams <- paramSearch[[gridId]][modelSettings$modelParamNames]
-
-      currentEstimatorSettings <-
-        fillEstimatorSettings(modelSettings$estimatorSettings,
-                              fitParams,
-                              paramSearch[[gridId]])
-      currentEstimatorSettings$modelType <- modelSettings$modelType
-      currentModelParams$catFeatures <- dataset$get_cat_features()$max()
-      currentModelParams$numFeatures <-
-        dataset$get_numerical_features()$len()
-      if (findLR) {
-        lrFinder <- createLRFinder(modelType = modelSettings$modelType,
-          modelParameters = currentModelParams,
-          estimatorSettings = currentEstimatorSettings
+        paramSearch[[gridId]],
+        collapse = " | "
+      ))
+      hyperparameterResults[[gridId]] <-
+        doCrossValidation(dataset,
+          labels = labels,
+          parameters = paramSearch[[gridId]],
+          modelSettings = modelSettings
         )
-        lr <- lrFinder$get_lr(dataset)
-        ParallelLogger::logInfo(paste0("Auto learning rate selected as: ", lr))
-        currentEstimatorSettings$learningRate <- lr
-      }
-
-      crossValidationResults <-
-        doCrossvalidation(dataset,
-                          labels = labels,
-                          modelSettings = currentModelParams,
-                          estimatorSettings = currentEstimatorSettings)
-      learnRates <- crossValidationResults$learnRates
-      prediction <- crossValidationResults$prediction
-
-      gridPerformance <-
-        PatientLevelPrediction::computeGridPerformance(prediction,
-                                                       paramSearch[[gridId]])
-      maxIndex <- which.max(unlist(sapply(learnRates, `[`, 2)))
-      gridSearchPredictons[[gridId]] <- list(
-        prediction = prediction,
-        param = paramSearch[[gridId]],
-        gridPerformance = gridPerformance
-      )
-      gridSearchPredictons[[gridId]]$gridPerformance$hyperSummary$learnRates <-
-        rep(list(unlist(learnRates[[maxIndex]]$LRs)),
-            nrow(gridSearchPredictons[[gridId]]$gridPerformance$hyperSummary))
-      gridSearchPredictons[[gridId]]$param$learnSchedule <-
-        learnRates[[maxIndex]]
-      # remove all predictions that are not the max performance
-      indexOfMax <-
-        which.max(unlist(lapply(gridSearchPredictons,
-                                function(x) x$gridPerformance$cvPerformance)))
-      for (i in seq_along(gridSearchPredictons)) {
-        if (!is.null(gridSearchPredictons[[i]]) && i != indexOfMax) {
-          gridSearchPredictons[[i]]$prediction <- list(NULL)
-        }
-      }
-      ParallelLogger::logInfo(paste0("Caching all grid search results and 
-                                     prediction for best combination ",
-                                     indexOfMax))
-      trainCache$saveGridSearchPredictions(gridSearchPredictons)
+     # remove all predictions that are not the max performance
+     hyperparameterResults <- trainCache$trimPerformance(hyperparameterResults) 
+     trainCache$saveGridSearchPredictions(hyperparameterResults)
     }
   }
-  paramGridSearch <- lapply(gridSearchPredictons, function(x) x$gridPerformance)
+  paramGridSearch <- lapply(hyperparameterResults,
+                            function(x) x$gridPerformance)
   # get best params
   indexOfMax <-
-    which.max(unlist(lapply(gridSearchPredictons,
-                            function(x) x$gridPerformance$cvPerformance)))
-  finalParam <- gridSearchPredictons[[indexOfMax]]$param
+    which.max(unlist(lapply(
+      hyperparameterResults,
+      function(x) x$gridPerformance$cvPerformance
+    )))
+  if (length(indexOfMax) == 0) {
+    stop("No hyperparameter combination has valid results")
+  }
+  finalParam <- hyperparameterResults[[indexOfMax]]$param
 
-  paramGridSearch <- lapply(gridSearchPredictons, function(x) x$gridPerformance)
+  paramGridSearch <- lapply(hyperparameterResults,
+                            function(x) x$gridPerformance)
 
   # get best CV prediction
-  cvPrediction <- gridSearchPredictons[[indexOfMax]]$prediction
+  cvPrediction <- hyperparameterResults[[indexOfMax]]$prediction
   cvPrediction$evaluationType <- "CV"
 
   ParallelLogger::logInfo("Training final model using optimal parameters")
-  # get the params
-  modelParams <- finalParam[modelSettings$modelParamNames]
-
-
-  # create the dir
+  trainPrediction <- trainFinalModel(dataset,
+                                     finalParam,
+                                     modelSettings,
+                                     labels)
+  prediction <- rbind(
+    trainPrediction$prediction,
+    cvPrediction
+  )
+  # remove fold index from predictions and remove cohortStartDate
+  prediction <- prediction %>%
+    dplyr::select(-"index")
+  prediction$cohortStartDate <- as.Date(prediction$cohortStartDate,
+    origin = "1970-01-01")
+  numericalIndex <- dataset$get_numerical_features()
+  
+  # save torch code here
   if (!dir.exists(file.path(modelLocation))) {
     dir.create(file.path(modelLocation), recursive = TRUE)
   }
-
-  modelParams$catFeatures <- dataset$get_cat_features()$max()
-  modelParams$numFeatures <- dataset$get_numerical_features()$len()
-
-
-  estimatorSettings <- fillEstimatorSettings(modelSettings$estimatorSettings,
-                                             fitParams,
-                                             finalParam)
-  estimatorSettings$learningRate <- finalParam$learnSchedule$LRs[[1]]
-  estimator <- createEstimator(modelType = modelSettings$modelType,
-                               modelParameters = modelParams,
-                               estimatorSettings = estimatorSettings)
-
-  numericalIndex <- dataset$get_numerical_features()
-  estimator$fit_whole_training_set(dataset, finalParam$learnSchedule$LRs)
-
-  ParallelLogger::logInfo("Calculating predictions on all train data...")
-  prediction <- predictDeepEstimator(
-    plpModel = estimator,
-    data = dataset,
-    cohort = labels
-  )
-  prediction$evaluationType <- "Train"
-
-  prediction <- rbind(
-    prediction,
-    cvPrediction
-  )
-  # modify prediction
-  prediction <- prediction %>%
-    dplyr::select(-"index")
-
-  prediction$cohortStartDate <- as.Date(prediction$cohortStartDate,
-                                        origin = "1970-01-01")
-
-
-  # save torch code here
-  estimator$save(modelLocation, "DeepEstimatorModel.pt")
+  trainPrediction$estimator$save(modelLocation, "DeepEstimatorModel.pt")
   return(
     list(
       estimator = modelLocation,
@@ -476,48 +465,159 @@ evalEstimatorSettings <- function(estimatorSettings) {
   estimatorSettings
 }
 
-createEstimator <- function(modelType,
-                            modelParameters,
+createEstimator <- function(modelParameters,
                             estimatorSettings) {
   path <- system.file("python", package = "DeepPatientLevelPrediction")
 
-  model <- reticulate::import_from_path(modelType, path = path)[[modelType]]
+  if (modelParameters$modelType == "Finetuner") {
+    estimatorSettings$finetune <- TRUE
+    plpModel <- PatientLevelPrediction::loadPlpModel(modelParameters$modelPath)
+    estimatorSettings$finetuneModelPath <-
+      normalizePath(file.path(plpModel$model, "DeepEstimatorModel.pt"))
+    modelParameters$modelType <-
+      plpModel$modelDesign$modelSettings$modelType
+  }
+
+  model <-
+    reticulate::import_from_path(modelParameters$modelType,
+                                 path = path)[[modelParameters$modelType]]
   estimator <- reticulate::import_from_path("Estimator", path = path)$Estimator
 
   modelParameters <- camelCaseToSnakeCaseNames(modelParameters)
   estimatorSettings <- camelCaseToSnakeCaseNames(estimatorSettings)
   estimatorSettings <- evalEstimatorSettings(estimatorSettings)
 
-  estimator <- estimator(model = model,
-                         model_parameters = modelParameters,
-                         estimator_settings = estimatorSettings)
+  estimator <- estimator(
+    model = model,
+    model_parameters = modelParameters,
+    estimator_settings = estimatorSettings
+  )
   return(estimator)
 }
 
-doCrossvalidation <- function(dataset,
+doCrossValidation <- function(dataset,
                               labels,
-                              modelSettings,
-                              estimatorSettings) {
+                              parameters,
+                              modelSettings
+                              ) {
+  crossValidationResults <-
+    tryCatch(doCrossValidationImpl(dataset,
+                                   labels,
+                                   parameters,
+                                   modelSettings),
+             error = function(e) {
+               if (inherits(e, "torch.cuda.OutOfMemoryError")) {
+                 ParallelLogger::logError(
+                   "Out of memory error during cross validation, 
+                   trying to continue with next hyperparameter combination"
+                 )
+                 crossValidationResults <- list()
+                 crossValidationResults$prediction <- labels
+                 crossValidationResults$prediction <- 
+                   cbind(crossValidationResults$prediction, value = NA)
+                 attr(crossValidationResults$prediction,
+                      "metaData")$modelType <- "binary"
+                 crossValidationResults$param <- parameters
+                 crossValidationResults$param$learnSchedule <- list(
+                   LRs = NA,
+                   bestEpoch = NA
+                 )
+                 nFolds <- max(labels$index)
+                 hyperSummary <- 
+                   data.frame(metric = rep("computeAuc", nFolds + 1),
+                              fold = c("CV", as.character(1:nFolds)),
+                              value = NA)
+                 hyperSummary <- cbind(hyperSummary, parameters)
+                 hyperSummary$learnRates <- NA
+                                    
+                 gridPerformance <- list(
+                   metric = "computeAuc",
+                   cvPerformance = NA,
+                   cvPerformancePerFold = rep(NA, nFolds),
+                   param = parameters,
+                   hyperSummary = hyperSummary
+                 )
+                crossValidationResults$gridPerformance <- gridPerformance
+                learnRates <- list()
+                for (i in 1:nFolds) {
+                  learnRates[[i]] <- list(
+                    LRs = NA,
+                    bestEpoch = NA
+                  )
+                }
+                crossValidationResults$learnRates <- learnRates
+                return(crossValidationResults)
+                } else {
+                  stop(e)
+                }
+              }
+            )
+  gridSearchPredictions <- list(
+    prediction = crossValidationResults$prediction,
+    param = parameters,
+    gridPerformance = crossValidationResults$gridPerformance
+  ) 
+  maxIndex <- which.max(unlist(sapply(crossValidationResults$learnRates, 
+                                      `[`, 2)))
+  if (length(maxIndex) != 0) {
+    gridSearchPredictions$gridPerformance$hyperSummary$learnRates <-
+      rep(
+        list(unlist(crossValidationResults$learnRates[[maxIndex]]$LRs)),
+        nrow(gridSearchPredictions$gridPerformance$hyperSummary)
+      )
+    gridSearchPredictions$param$learnSchedule <-
+      crossValidationResults$learnRates[[maxIndex]]
+  }
+  return(gridSearchPredictions)
+}
+
+doCrossValidationImpl <- function(dataset,
+                                  labels,
+                                  parameters,
+                                  modelSettings) {
+  fitParams <- names(parameters)[grepl(
+    "^estimator",
+    names(parameters)
+  )]
+  currentModelParams <- parameters[modelSettings$modelParamNames]
+  attr(currentModelParams, "metaData")$names <-
+    modelSettings$modelParamNameCH
+  currentModelParams$modelType <- modelSettings$modelType
+  currentEstimatorSettings <-
+    fillEstimatorSettings(modelSettings$estimatorSettings,
+                          fitParams,
+                          parameters)
+  currentModelParams$catFeatures <- dataset$get_cat_features()$max()
+  currentModelParams$numFeatures <- dataset$get_numerical_features()$len()
+  if (currentEstimatorSettings$findLR) {
+    lr <- getLR(currentModelParams, currentEstimatorSettings, dataset)
+    ParallelLogger::logInfo(paste0("Auto learning rate selected as: ", lr))
+    currentEstimatorSettings$learningRate <- lr
+  }
+
   fold <- labels$index
   ParallelLogger::logInfo(paste0("Max fold: ", max(fold)))
   learnRates <- list()
   prediction <- NULL
+  path <- system.file("python", package = "DeepPatientLevelPrediction")
+  fit_estimator <- reticulate::import_from_path("Estimator",
+                                                path = path)$fit_estimator
   for (i in 1:max(fold)) {
     ParallelLogger::logInfo(paste0("Fold ", i))
 
     # -1 for python 0-based indexing
     trainDataset <- torch$utils$data$Subset(dataset,
-                                            indices =
-                                              as.integer(which(fold != i) - 1))
+      indices =
+        as.integer(which(fold != i) - 1)
+    )
 
     # -1 for python 0-based indexing
     testDataset <- torch$utils$data$Subset(dataset,
                                            indices =
                                              as.integer(which(fold == i) - 1))
-    estimator <- createEstimator(modelType = estimatorSettings$modelType,
-                                 modelParameters = modelSettings,
-                                 estimatorSettings = estimatorSettings)
-    estimator$fit(trainDataset, testDataset)
+    estimator <- createEstimator(modelParameters = currentModelParams,
+                                 estimatorSettings = currentEstimatorSettings)
+    fit_estimator(estimator, trainDataset, testDataset)
 
     ParallelLogger::logInfo("Calculating predictions on left out fold set...")
 
@@ -534,10 +634,15 @@ doCrossvalidation <- function(dataset,
       bestEpoch = estimator$best_epoch
     )
   }
-  return(results = list(prediction = prediction,
-                        learnRates = learnRates))
-
+  gridPerformance <- PatientLevelPrediction::computeGridPerformance(prediction,
+                                                                    parameters)
+  return(results = list(
+    prediction = prediction,
+    learnRates = learnRates,
+    gridPerformance = gridPerformance
+  ))
 }
+
 
 extractParamsToTune <- function(estimatorSettings) {
   paramsToTune <- list()
@@ -556,4 +661,35 @@ extractParamsToTune <- function(estimatorSettings) {
     }
   }
   return(paramsToTune)
+}
+
+trainFinalModel <- function(dataset, finalParam, modelSettings, labels) {
+    # get the params
+    modelParams <- finalParam[modelSettings$modelParamNames]
+  
+    fitParams <- names(finalParam)[grepl("^estimator", names(finalParam))]
+    
+    modelParams$catFeatures <- dataset$get_cat_features()$max()
+    modelParams$numFeatures <- dataset$get_numerical_features()$len()
+    modelParams$modelType <- modelSettings$modelType
+  
+    estimatorSettings <- fillEstimatorSettings(
+      modelSettings$estimatorSettings,
+      fitParams,
+      finalParam
+    )
+    estimatorSettings$learningRate <- finalParam$learnSchedule$LRs[[1]]
+    estimator <- createEstimator(modelParameters = modelParams,
+                                 estimatorSettings = estimatorSettings)
+    estimator$fit_whole_training_set(dataset, finalParam$learnSchedule$LRs)
+  
+    ParallelLogger::logInfo("Calculating predictions on all train data...")
+    prediction <- predictDeepEstimator(
+      plpModel = estimator,
+      data = dataset,
+      cohort = labels
+    )
+    prediction$evaluationType <- "Train"
+    return(list(prediction = prediction, 
+                estimator = estimator))
 }