+++ inst/python/LrFinder.py
@@ -1,3 +1,4 @@
+from os import walk
 import random
 
 import torch
@@ -5,6 +6,7 @@ from torch.optim.lr_scheduler import _LRScheduler
 from tqdm import tqdm
 
 from Estimator import batch_to_device
+from gpu_memory_cleanup import memory_cleanup
 
 
 class ExponentialSchedulerPerBatch(_LRScheduler):
@@ -19,71 +21,70 @@ class ExponentialSchedulerPerBatch(_LRScheduler):
 
 
 class LrFinder:
-    def __init__(self, model, model_parameters, estimator_settings, lr_settings):
+    def __init__(self, estimator, lr_settings=None):
+        self.first_batch = None
         if lr_settings is None:
             lr_settings = {}
-        min_lr = lr_settings.get("min_lr", 1e-7)
-        max_lr = lr_settings.get("max_lr", 1)
-        num_lr = lr_settings.get("num_lr", 100)
-        smooth = lr_settings.get("smooth", 0.05)
-        divergence_threshold = lr_settings.get("divergence_threshold", 4)
-        torch.manual_seed(seed=estimator_settings["seed"])
-        self.seed = estimator_settings["seed"]
-        self.model = model(**model_parameters)
-        if callable(estimator_settings["device"]):
-            self.device = estimator_settings["device"]()
+        self.min_lr = lr_settings.get("min_lr", 1e-7)
+        self.max_lr = lr_settings.get("max_lr", 1)
+        self.num_lr = lr_settings.get("num_lr", 100)
+        self.smooth = lr_settings.get("smooth", 0.05)
+        self.divergence_threshold = lr_settings.get("divergence_threshold", 4)
+        torch.manual_seed(seed=estimator.seed)
+        self.seed = estimator.seed
+        if self.num_lr < 20:
+            self.min_factor = 0
         else:
-            self.device = estimator_settings["device"]
-        self.model.to(device=self.device)
-        self.min_lr = min_lr
-        self.max_lr = max_lr
-        self.num_lr = num_lr
-        self.smooth = smooth
-        self.divergence_threshold = divergence_threshold
+            self.min_factor = 5
 
-        self.optimizer = estimator_settings["optimizer"](
-            params=self.model.parameters(), lr=self.min_lr
+        for group in estimator.optimizer.param_groups:
+            group['lr'] = self.min_lr
+
+        estimator.scheduler = ExponentialSchedulerPerBatch(
+            estimator.optimizer, self.max_lr, self.num_lr
         )
-
-        self.scheduler = ExponentialSchedulerPerBatch(
-            self.optimizer, self.max_lr, self.num_lr
-        )
-
-        self.criterion = estimator_settings["criterion"]()
-        self.batch_size = int(estimator_settings["batch_size"])
+        if estimator.accumulation_steps > 1:
+            self.accumulation_steps = estimator.accumulation_steps
+        else:
+            self.accumulation_steps = 1
+        self.estimator = estimator
         self.losses = None
         self.loss_index = None
 
     def get_lr(self, dataset):
+        if len(dataset) < self.estimator.batch_size:
+            self.estimator.batch_size = len(dataset)
         batch_index = torch.arange(0, len(dataset), 1).tolist()
         random.seed(self.seed)
         losses = torch.empty(size=(self.num_lr,), dtype=torch.float)
         lrs = torch.empty(size=(self.num_lr,), dtype=torch.float)
+        self.estimator.optimizer.zero_grad()
+        best_loss = float("inf")
         for i in tqdm(range(self.num_lr)):
-            self.optimizer.zero_grad()
-            random_batch = random.sample(batch_index, self.batch_size)
-            batch = dataset[random_batch]
-            batch = batch_to_device(batch, self.device)
+            loss_value = 0
+            random_batch = random.sample(batch_index, self.estimator.batch_size)
+            for j in range(self.accumulation_steps):
+                batch = dataset[random_batch[j * self.estimator.sub_batch_size : (j + 1) * self.estimator.sub_batch_size]]
+                batch = batch_to_device(batch, self.estimator.device)
 
-            out = self.model(batch[0])
-            loss = self.criterion(out, batch[1])
+                out = self.estimator.model(batch[0])
+                loss = self.estimator.criterion(out, batch[1])
+                loss.backward()
+                loss_value += loss.item()
+            loss_value = loss_value / self.accumulation_steps
             if self.smooth is not None and i != 0:
                 losses[i] = (
-                    self.smooth * loss.item() + (1 - self.smooth) * losses[i - 1]
+                    self.smooth * loss_value + (1 - self.smooth) * losses[i - 1]
                 )
             else:
-                losses[i] = loss.item()
-            lrs[i] = self.optimizer.param_groups[0]["lr"]
+                losses[i] = loss_value
+            lrs[i] = self.estimator.optimizer.param_groups[0]["lr"]
 
-            loss.backward()
-            self.optimizer.step()
-            self.scheduler.step()
+            self.estimator.optimizer.step()
+            self.estimator.scheduler.step()
 
-            if i == 0:
+            if losses[i] < best_loss:
                 best_loss = losses[i]
-            else:
-                if losses[i] < best_loss:
-                    best_loss = losses[i]
 
             if losses[i] > (self.divergence_threshold * best_loss):
                 print(
@@ -93,12 +94,29 @@ class LrFinder:
 
         # find LR where gradient is highest but before global minimum is reached
         # I added -5 to make sure it is not still in the minimum
-        global_minimum = torch.argmin(losses)
-        gradient = torch.diff(losses[: (global_minimum - 5) + 1])
-        smallest_gradient = torch.argmin(gradient)
+        global_minimum = torch.argmin(losses[:i + 1])
+        if global_minimum == 0:
+            biggest_gradient = 0
+        else:
+            gradient = torch.diff(losses[: (global_minimum - self.min_factor) + 1])
+            biggest_gradient = torch.argmax(gradient)
 
-        suggested_lr = lrs[smallest_gradient]
-        self.losses = losses
-        self.loss_index = smallest_gradient
+        suggested_lr = lrs[biggest_gradient]
+        self.losses = losses[:i]
+        self.loss_index = biggest_gradient
         self.lrs = lrs
         return suggested_lr.item()
+
+
+def get_lr(estimator,
+           dataset,
+           lr_settings):
+    try:
+        lr_finder = LrFinder(estimator, lr_settings=lr_settings)
+        lr = lr_finder.get_lr(dataset)
+    except torch.cuda.OutOfMemoryError as e:
+        memory_cleanup()
+        raise e
+    return lr
+
+